+++ classes/WpMatomo/Admin/Dashboard.php
@@ -10,9 +10,11 @@
 namespace WpMatomo\Admin;
 
 use WpMatomo\Capabilities;
+use WpMatomo\Logger;
 use WpMatomo\Report\Dates;
 use WpMatomo\Report\Metadata;
 use WpMatomo\Report\Renderer;
+use WpMatomo\Uninstaller;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // if accessed directly
@@ -29,22 +31,31 @@ class Dashboard {
 	public function add_dashboard_widgets()
 	{
 		$widgets = $this->get_widgets();
-		if (!empty($widgets) && is_array($widgets)) {
+		if (!empty($widgets) && is_array($widgets) && current_user_can(Capabilities::KEY_VIEW)) {
 			foreach ($widgets as $widget) {
-				$widget_meta = $this->is_valid_widget($widget['unique_id'], $widget['date']);
-				if (!empty($widget_meta)) {
-					$id = 'matomo_dashboard_widget_' . $widget['unique_id'] .  '_' . $widget['date'];
 
-					$title = esc_html($widget_meta['report']['name'] . ' - ' . $widget_meta['date'] . ' - Matomo');
-					wp_add_dashboard_widget( $id, $title, function () use ($widget) {
-						$renderer = new Renderer();
-						echo $renderer->show_report(array(
-							'unique_id'   => $widget['unique_id'],
-							'report_date' => $widget['date'],
-							'limit'       => 10,
-						));
-					});
-				}
+			    try {
+
+                    $widget_meta = $this->is_valid_widget($widget['unique_id'], $widget['date']);
+                    if (!empty($widget_meta['report']['name'])) {
+                        $id = 'matomo_dashboard_widget_' . $widget['unique_id'] .  '_' . $widget['date'];
+
+                        $title = $widget_meta['report']['name'] . ' - ' . $widget_meta['date'] . ' - Matomo';
+                        wp_add_dashboard_widget( $id, esc_html($title), function () use ($widget) {
+                            $renderer = new Renderer();
+                            echo $renderer->show_report(array(
+                                'unique_id'   => $widget['unique_id'],
+                                'report_date' => $widget['date'],
+                                'limit'       => 10,
+                            ));
+                        });
+                    }
+                } catch (\Exception $e) {
+			        // dont want to break dashboard if there is any issue with matomo ... eg in case bootstrap fails
+                    // or is reinstalled but matomo not yet fully installed etc
+			        $logger = new Logger();
+			        $logger->log(sprintf('Failed to add Matomo widget %s to dashboard: %s', wp_json_encode($widget), $e->getMessage()));
+                }
 			}
 		}
 	}
@@ -54,8 +65,10 @@ class Dashboard {
 		if (empty($unique_id) || empty($date)) {
 			return false;
 		}
+
 		$metadata = new Metadata();
 		$report = $metadata->find_report_by_unique_id( $unique_id );
+
 		if (empty($report)) {
 			return false;
 		}
@@ -87,7 +100,7 @@ class Dashboard {
 		foreach ($widgets as $index => $widget) {
 			if ($widget['unique_id'] === $report_unique_id && $widget['date'] === $report_date) {
 				unset($widgets[$index]);
-				$this->set_widgets($widgets);
+				$this->set_widgets(array_values($widgets));
 				return;
 			}
 		}
@@ -111,6 +124,6 @@ class Dashboard {
 	}
 
 	public function uninstall() {
-		delete_user_meta(get_current_user_id(), self::DASHBOARD_USER_OPTION);
+	    Uninstaller::uninstall_user_meta(self::DASHBOARD_USER_OPTION);
 	}
 }