+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Support UTF8MB4 character set in tracking
 * Improve compatibility with some plugins fixing some archiving issues
 * Improve tracking settings for multi sites
+* Add widgets to dashboard from the summary page
 * Show a blog selector in the reporting page when using multi site
 * Fix super admins weren't always synced in multisite mode
 * By default delete all data on plugin uninstallation unless configured differently